Implement simple test harness running printer via remoting
